pyproject.toml
[project]
name = "graph-rag"
version = "0.1.0"
description = "Graph RAG with NebulaGraph, configurable vector stores, embedding models and LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-restx>=1.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "nebula3-python>=3.6.0",
    "qdrant-client>=1.7.0",
    "openai>=1.10.0",
    "httpx>=0.26.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    # 文档解析
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "chardet>=5.2.0",
    # 文本分块
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[tool.uv.sources]
# 可在此添加自定义源，例如:
# some-package = { git = "https://github.com/user/repo" }
